src/functions/predict_function/requirements.txt
azure-functions==1.20.0
azure-storage-blob==12.22.0
requests==2.32.3
